+++ packages/esp32/loramint/__init__.py
@@ -1,0 +1,13 @@
+"""
+LoRaMINT - MicroPython library for LoRaMINT sensor nodes on an ESP32 with a
+Dragino LA66 LoRaWAN module.
+
+    from loramint import LoRaMINT, MintValue
+"""
+
+from .loramint import LoRaMINT
+from .mintvalue import MintValue
+
+__version__ = "0.1.0"
+
+__all__ = ["LoRaMINT", "MintValue"]